Clinical Decision Support/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.15.0
openai>=0.27.0
fastapi>=0.68.0
python-dotenv>=0.19.0
pydantic>=1.9.0
python-multipart>=0.0.5